+++ src/main/java/com/sandymandy/pleasurecraft/PleasureCraftClient.java
@@ -49,7 +49,8 @@ public class PleasureCraftClient implements ClientModInitializer {
         EntityRendererRegistry.register(GirlRegistry.MIKA, MikaRenderer::new);
         EntityRendererRegistry.register(GirlRegistry.MOMO, MomoRenderer::new);
         EntityRendererRegistry.register(GirlRegistry.SLIME, SlimeRenderer::new);
-		EntityRendererRegistry.register(GirlRegistry.KOBOLD, KoboldRenderer::new);
+        EntityRendererRegistry.register(GirlRegistry.KOBOLD, KoboldRenderer::new);
+        EntityRendererRegistry.register(GirlRegistry.COPPIE, CoppieRenderer::new);
         EntityRendererRegistry.register(GirlRegistry.CUSTOM_GIRL, CustomGirlRenderer::new);
         PleasureCraftKeybinds.register();
         PleasureCraftClientPackets.registerS2CPackets();