+++ src/main/java/com/sandymandy/pleasurecraft/client/models/CoppieModel.java
@@ -1,0 +1,6 @@
+package com.sandymandy.pleasurecraft.client.models;
+
+import com.sandymandy.pleasurecraft.entity.girls.CoppieEntity;
+
+public class CoppieModel extends AbstractGirlModel<CoppieEntity> {
+}