+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added a new masturbation scene to lucy
 - Made the pregnancy progress save
 - Updated the girl customize screen so it acts like the kobold one
+- Added a new item that adds girls to settlements 
+- Added a new girl based on the copper golem called Coppie
 
 ## v0.6.1
 - Fixed the issue where girls in villagers would have tiny breasts